refactor: break down large stack.rs file into smaller focused modules
- Split stack.rs (1,518 lines) into types.rs and stack.rs
- Move tests from inline #[cfg(test)] to dedicated test file
- Create types.rs with all data structures and enums
- Reduce stack.rs to core implementation only (~307 lines)
- Update module declarations and fix import references
- Maintain backward compatibility through public re-exports
- All 34 iterator stack tests pass

Files changed:
- src/transform/iterator_stack/types.rs (new)
- src/transform/iterator_stack/stack.rs (refactored)
- tests/unit/iterator_stack_tests.rs (new)
- Updated execution engine imports
- Updated module declarations
