+++ tests/unit/mod.rs
@@ -13,3 +13,4 @@ pub mod transform_utils_helper_tests;
 pub mod declarative_transform_tests;
 pub mod hashrange_schema_tests;
 pub mod schema_declarative_transform_interpretation_tests;
+pub mod iterator_stack_tests;
